Заканчивает модуляцию
